decouple cycle-import
